+++ gazebo/Server.cc
@@ -101,6 +101,9 @@ namespace gazebo
 
     /// \brief Save argv for access by system plugins.
     char **systemPluginsArgv;
+
+    /// \brief Set whether to lockstep physics and rendering
+    bool lockstep = false;
   };
 }
 
@@ -150,6 +153,7 @@ bool Server::ParseArgs(int _argc, char **_argv)
     ("verbose", "Increase the messages written to the terminal.")
     ("help,h", "Produce this help message.")
     ("pause,u", "Start the server in a paused state.")
+    ("lockstep", "Lockstep simulation so sensor update rates are respected.")
     ("physics,e", po::value<std::string>(),
      "Specify a physics engine (ode|bullet|dart|simbody).")
     ("play,p", po::value<std::string>(), "Play a log file.")
@@ -282,6 +286,12 @@ bool Server::ParseArgs(int _argc, char **_argv)
         this->dataPtr->vm["iters"].as<unsigned int>() << "]\n";
     }
   }
+
+  if (this->dataPtr->vm.count("lockstep"))
+  {
+    this->dataPtr->lockstep = true;
+  }
+  rendering::set_lockstep_enabled(this->dataPtr->lockstep);
 
   if (!this->PreLoad())
   {
@@ -498,7 +508,11 @@ bool Server::LoadImpl(sdf::ElementPtr _elem,
       << " seconds for namespaces. Giving up.\n";
   }
 
-  physics::init_worlds(rendering::update_scene_poses);
+  if (this->dataPtr->lockstep)
+    physics::init_worlds(rendering::update_scene_poses);
+  else
+    physics::init_worlds(nullptr);
+
   this->dataPtr->stop = false;
 
   return true;
@@ -590,6 +604,12 @@ void Server::Run()
   {
     IGN_PROFILE("Server::Run");
     IGN_PROFILE_BEGIN("ProcessControlMsgs");
+    if (this->dataPtr->lockstep)
+      rendering::wait_for_render_request("", 0.100);
+    // bool ret = rendering::wait_for_render_request("", 0.100);
+    // if (ret == false)
+    //   gzerr << "time out reached!" << std::endl;
+
     this->ProcessControlMsgs();
     IGN_PROFILE_END();
 
@@ -606,7 +626,8 @@ void Server::Run()
       IGN_PROFILE_END();
     }
 
-    common::Time::MSleep(1);
+    if (!this->dataPtr->lockstep)
+      common::Time::MSleep(1);
   }
 
   // Shutdown gazebo