+++ gazebo/rendering/RenderingIface.cc
@@ -24,6 +24,8 @@
 
 using namespace gazebo;
 
+bool g_lockstep = false;
+
 //////////////////////////////////////////////////
 bool rendering::load()
 {
@@ -110,6 +112,21 @@ void rendering::remove_scene(const std::string &_name)
 }
 
 //////////////////////////////////////////////////
+bool rendering::wait_for_render_request(const std::string &_name,
+                                        const double _timeoutsec)
+{
+  ScenePtr scene = get_scene(_name);
+
+  if (!scene)
+  {
+    common::Time::NSleep(_timeoutsec * 1e9);
+    return false;
+  }
+
+  return scene->WaitForRenderRequest(_timeoutsec);
+}
+
+//////////////////////////////////////////////////
 void rendering::update_scene_poses(const std::string &_name,
                                    const msgs::PosesStamped &_msg)
 {
@@ -118,4 +135,16 @@ void rendering::update_scene_poses(const std::string &_name,
   {
     scene->UpdatePoses(_msg);
   }
+}
+
+//////////////////////////////////////////////////
+void rendering::set_lockstep_enabled(bool _enable)
+{
+  g_lockstep = _enable;
+}
+
+//////////////////////////////////////////////////
+bool rendering::lockstep_enabled()
+{
+  return g_lockstep;
 }