+++ gazebo/sensors/CameraSensor.cc
@@ -26,6 +26,7 @@
 
 #include "gazebo/msgs/msgs.hh"
 
+#include "gazebo/physics/PhysicsEngine.hh"
 #include "gazebo/physics/World.hh"
 
 #include "gazebo/rendering/Camera.hh"
@@ -66,6 +67,13 @@ CameraSensor::~CameraSensor()
 void CameraSensor::Load(const std::string &_worldName, sdf::ElementPtr _sdf)
 {
   Sensor::Load(_worldName, _sdf);
+  // useStrictRate is set in Sensor::Load()
+  if (this->useStrictRate)
+  {
+    this->connections.push_back(
+        event::Events::ConnectPreRenderEnded(
+          boost::bind(&CameraSensor::PrerenderEnded, this)));
+  }
 }
 
 //////////////////////////////////////////////////
@@ -92,6 +100,14 @@ std::string CameraSensor::TopicIgn() const
 void CameraSensor::Load(const std::string &_worldName)
 {
   Sensor::Load(_worldName);
+  // useStrictRate is set in Sensor::Load()
+  if (this->useStrictRate)
+  {
+    this->connections.push_back(
+        event::Events::ConnectPreRenderEnded(
+          boost::bind(&CameraSensor::PrerenderEnded, this)));
+  }
+
   this->imagePub = this->node->Advertise<msgs::ImageStamped>(this->Topic(), 50);
 
   ignition::transport::AdvertiseMessageOptions opts;
@@ -194,13 +210,66 @@ void CameraSensor::Fini()
 //////////////////////////////////////////////////
 void CameraSensor::SetActive(bool _value)
 {
+  // If this sensor is reactivated
+  if (this->useStrictRate && _value && !this->IsActive())
+  {
+    // the next rendering time must be reset to ensure it is properly
+    // computed by CameraSensor::NeedsUpdate.
+    this->dataPtr->nextRenderingTime = std::numeric_limits<double>::quiet_NaN();
+  }
   Sensor::SetActive(_value);
 }
 
 //////////////////////////////////////////////////
 bool CameraSensor::NeedsUpdate()
 {
-  return Sensor::NeedsUpdate();
+  if (this->useStrictRate)
+  {
+    double simTime;
+    if (this->scene)
+      simTime = this->scene->SimTime().Double();
+    else
+      simTime = this->world->SimTime().Double();
+
+    if (simTime < this->lastMeasurementTime.Double())
+    {
+      // Rendering sensors also set the lastMeasurementTime variable in Render()
+      // and lastUpdateTime in Sensor::Update based on Scene::SimTime() which
+      // could be outdated when the world is reset. In this case reset
+      // the variables back to 0.
+      this->ResetLastUpdateTime();
+      return false;
+    }
+
+    double dt = this->world->Physics()->GetMaxStepSize();
+
+    // If next rendering time is not set yet
+    if (std::isnan(this->dataPtr->nextRenderingTime))
+    {
+      if (this->updatePeriod == 0
+          || (simTime > 0.0 &&
+          std::abs(std::fmod(simTime, this->updatePeriod.Double())) < dt))
+      {
+        this->dataPtr->nextRenderingTime = simTime;
+        return true;
+      }
+      else
+      {
+        return false;
+      }
+    }
+
+    if (simTime > this->dataPtr->nextRenderingTime + dt)
+      return true;
+
+    // Trigger on the tick the closest from the targeted rendering time
+    return (ignition::math::lessOrNearEqual(
+          std::abs(simTime - this->dataPtr->nextRenderingTime), dt / 2.0));
+  }
+  else
+  {
+    return Sensor::NeedsUpdate();
+  }
 }
 
 //////////////////////////////////////////////////
@@ -210,16 +279,49 @@ void CameraSensor::Update(bool _force)
 }
 
 //////////////////////////////////////////////////
+void CameraSensor::PrerenderEnded()
+{
+  if (this->useStrictRate && this->camera && this->IsActive() &&
+      this->NeedsUpdate())
+  {
+    // compute next rendering time, take care of the case where period is zero.
+    double dt;
+    if (this->updatePeriod <= 0.0)
+      dt = this->world->Physics()->GetMaxStepSize();
+    else
+      dt = this->updatePeriod.Double();
+    this->dataPtr->nextRenderingTime += dt;
+
+    this->dataPtr->renderNeeded = true;
+    this->lastMeasurementTime = this->scene->SimTime();
+  }
+}
+
+//////////////////////////////////////////////////
 void CameraSensor::Render()
 {
-  if (!this->camera || !this->IsActive() || !this->NeedsUpdate())
-    return;
+  if (this->useStrictRate)
+  {
+    if (!this->dataPtr->renderNeeded)
+      return;
 
-  // Update all the cameras
-  this->camera->Render();
+    // Update all the cameras
+    this->camera->Render();
 
-  this->dataPtr->rendered = true;
-  this->lastMeasurementTime = this->scene->SimTime();
+    this->dataPtr->rendered = true;
+    this->dataPtr->renderNeeded = false;
+  }
+  else
+  {
+    if (!this->camera || !this->IsActive() || !this->NeedsUpdate())
+      return;
+
+    // Update all the cameras
+    this->camera->Render();
+
+    this->dataPtr->rendered = true;
+    this->lastMeasurementTime = this->scene->SimTime();
+  }
 }
 
 //////////////////////////////////////////////////
@@ -367,11 +469,23 @@ void CameraSensor::SetRendered(const bool _value)
 //////////////////////////////////////////////////
 double CameraSensor::NextRequiredTimestamp() const
 {
-  return Sensor::NextRequiredTimestamp();
+  if (this->useStrictRate)
+  {
+    if (!ignition::math::equal(this->updatePeriod.Double(), 0.0))
+      return this->dataPtr->nextRenderingTime;
+    else
+      return std::numeric_limits<double>::quiet_NaN();
+  }
+  else
+  {
+    return Sensor::NextRequiredTimestamp();
+  }
 }
 
 //////////////////////////////////////////////////
 void CameraSensor::ResetLastUpdateTime()
 {
   Sensor::ResetLastUpdateTime();
+  if (this->useStrictRate)
+    this->dataPtr->nextRenderingTime = std::numeric_limits<double>::quiet_NaN();
 }