+++ gazebo/sensors/CameraSensorPrivate.hh
@@ -17,6 +17,8 @@
 #ifndef GAZEBO_SENSORS_CAMERASENSOR_PRIVATE_HH_
 #define GAZEBO_SENSORS_CAMERASENSOR_PRIVATE_HH_
 
+#include <limits>
+
 namespace gazebo
 {
   namespace sensors
@@ -27,6 +29,13 @@ namespace gazebo
     {
       /// \brief True if the sensor was rendered.
       public: bool rendered = false;
+
+      /// \brief True if the sensor needs a rendering
+      public: bool renderNeeded = false;
+
+      /// \brief Timestamp of the forthcoming rendering
+      public: double nextRenderingTime
+                           = std::numeric_limits<double>::quiet_NaN();
     };
   }
 }